Add ranks (#354)
* update cron jobs and leaderboard calculation

* feat: Add rank to users table

* feat: add to cron job
