+++ vercel.json
@@ -2,7 +2,15 @@
   "crons": [
     {
       "path": "/api/cron/leaderboard_update",
-      "schedule": "0 9 * * *"
+      "schedule": "5 9 * * 2"
+    },
+    {
+      "path": "/api/cron/stats_recalculation",
+      "schedule": "0 9 * * 2"
+    },
+    {
+      "path": "/api/cron/rank_update",
+      "schedule": "10 9 * * 2"
     }
   ],
   "functions": {
